ln: avoid recursive dependencies, make new lnutil
